renamed Dashboard to HostDashboard
